+++ django_spire/ai/chat/urls/json_urls.py
@@ -8,5 +8,5 @@ app_name = 'json'
 urlpatterns = [
     path('<int:pk>/delete', json_views.delete_chat_view, name='delete'),
     path('<int:pk>/rename', json_views.rename_chat_view, name='rename'),
-    path('workflow_process', json_views.chat_workflow_process_view, name='workflow_process'),
+    path('workflow_process', json_views.chat_workflow_view, name='workflow_process'),
 ]