+++ django_spire/ai/context/apps.py
@@ -9,14 +9,6 @@ class AiContextConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'django_spire.ai.context'
     label = 'django_spire_ai_context'
-    MODEL_PERMISSIONS = (
-        {
-            'name': 'ai_context',
-            'verbose_name': 'AI Context',
-            'model_class_path': 'django_spire.ai.context.models.Context',
-            'is_proxy_model': False,
-        },
-    )
 
     REQUIRED_APPS = ('django_spire_ai',)