+++ django_spire/knowledge/intelligence/workflows/knowledge_workflow.py
@@ -1,16 +1,30 @@
 from __future__ import annotations
 
-from django_spire.knowledge.intelligence.bots.entry_search_llm_bot import EntrySearchLlmBot
+from dandy.llm.request.message import MessageHistory
+from django.core.handlers.wsgi import WSGIRequest
+
+from django_spire.knowledge.intelligence.bots.entry_search_llm_bot import (
+    EntrySearchLlmBot,
+)
 from django_spire.knowledge.intelligence.intel.collection_intel import CollectionIntel
-from django_spire.knowledge.intelligence.intel.entry_intel import EntriesIntel, EntryIntel
-from django_spire.knowledge.intelligence.intel.message_intel import KnowledgeMessageIntel
-from django_spire.knowledge.intelligence.maps.collection_map import get_collection_map_class
+from django_spire.knowledge.intelligence.intel.entry_intel import (
+    EntriesIntel,
+    EntryIntel,
+)
+from django_spire.knowledge.intelligence.intel.message_intel import (
+    KnowledgeMessageIntel,
+)
+from django_spire.knowledge.intelligence.maps.collection_map import (
+    get_collection_map_class,
+)
 from django_spire.knowledge.intelligence.maps.entry_map import get_entry_map_class
 
 
 class KnowledgeWorkflow:
     @classmethod
-    def process(cls, user_input: str) -> KnowledgeMessageIntel:
+    def process(
+        cls, request: WSGIRequest, user_input: str, message_history: MessageHistory
+    ) -> KnowledgeMessageIntel:
         CollectionMap = get_collection_map_class()
         collections = CollectionMap().process(user_input).values
 
@@ -43,11 +57,8 @@ class KnowledgeWorkflow:
         entries_intel = EntriesIntel(
             entry_intel_list=[
                 EntryIntel(
-                    body=entry_search_bot.process(
-                        user_input=user_input,
-                        entry=entry
-                    ),
-                    collection_intel=CollectionIntel(name=entry.collection.name)
+                    body=entry_search_bot.process(user_input=user_input, entry=entry),
+                    collection_intel=CollectionIntel(name=entry.collection.name),
                 )
                 for entry in entries
             ]