+++ test_project/base_settings.py
@@ -46,9 +46,9 @@ TWILIO_PHONE_NUMBER = os.getenv('TWILIO_PHONE_NUMBER')
 TWILIO_SMS_BATCH_SIZE = 100
 
 # AI Chat Config
-AI_CHAT_WORKFLOW_NAME = 'Rubber Ducky'
-AI_CHAT_WORKFLOW_CLASS = 'test_project.apps.ai.chat.intelligence.workflows.chat_workflow.ChatWorkflow'
-AI_SMS_CONVERSATION_WORKFLOW_CLASS = 'test_project.apps.ai.sms.intelligence.workflows.sms_conversation_workflow.SmsConversationWorkflow'
+AI_PERSONA_NAME = 'Rubber Ducky'
+AI_CHAT_CALLABLE = 'test_project.apps.ai.chat.intelligence.workflows.chat_workflow.chat_workflow'
+AI_SMS_CONVERSATION_CALLABLE = 'test_project.apps.ai.sms.intelligence.workflows.sms_conversation_workflow.sms_conversation_workflow'
 
 ORGANIZATION_NAME = 'Synergistic Energistic Inc.'
 ORGANIZATION_DESCRIPTION = 'Marketing organization focused on the word synergy and energetically saying it as much as possible.'
@@ -79,6 +79,7 @@ INSTALLED_APPS = [
 INSTALLED_APPS += [
     'django_spire.ai',
     'django_spire.ai.chat',
+    'django_spire.ai.context',
     'django_spire.ai.sms',
 
     'django_spire.auth',